refactor: colocate rule contracts with database implementation
